refactor: wire up new slice
