+++ static/js/redux/constants/actionTypes.jsx
@@ -29,10 +29,6 @@ export const REQUEST_FRIENDS = "REQUEST_FRIENDS";
 // INIT
 export const INIT_STATE = "INIT_STATE";
 
-// INTEGRATIONS
-export const TOGGLE_INTEGRATION_MODAL = "TOGGLE_INTEGRATION_MODAL";
-export const OPEN_INTEGRATION_MODAL = "OPEN_INTEGRATION_MODAL";
-
 // NOTIFICATIONS
 export const TOKEN_REGISTERED = "TOKEN_REGISTERED";
 export const UNREGISTER_TOKEN = "UNREGISTER_TOKEN";
